+++ README.md
@@ -1,7 +1,7 @@
 # ExplainaBoard Client
 
-This is a command line and API client that makes it easy for you to upload systems to
-[ExplainaBoard](https://explainaboard.inspiredco.ai).
+This is a command line and API client that makes it easy for you to evaluate systems
+using [ExplainaBoard](https://explainaboard.inspiredco.ai).
 
 ## Preparation
 
@@ -29,15 +29,18 @@ export EB_API_KEY="[your API key]"
 
 ## Usage
 
-### Uploading/Browsing/Deleting Systems from the Command Line
+### Evaluating/Browsing/Deleting Systems from the Command Line
 
-**Uploading Systems:** The most common usage of this client will probably be to upload 
-systems. You can do that from the command line. If you are using a pre-existing dataset 
-viewable from the [ExplainaBoard datasets](https://explainaboard.inspiredco.ai/datasets)
+**Evaluating Systems:** The most common usage of this client will probably be to
+evaluate systems on the ExplainaBoard server. You can do that from the
+command line.
+
+If you are using a pre-existing dataset viewable from the
+[ExplainaBoard datasets](https://explainaboard.inspiredco.ai/datasets)
 page then you can use something like the following command:
 
 ```
-python -m explainaboard_client.cli.upload_system \
+python -m explainaboard_client.cli.evaluate_system \
   --email $EB_EMAIL --api_key $EB_API_KEY \
   --task [TASK_ID] \
   --system_name [MODEL_NAME] \
@@ -50,9 +53,9 @@ python -m explainaboard_client.cli.upload_system \
 You will need to fill in all the settings appropriately, for example:
 * `[TASK_ID]` is the ID of the task you want to perform. A full list is [here](https://github.com/neulab/explainaboard_web/blob/main/backend/src/impl/tasks.py).
 * `[MODEL_NAME]` is whatever name you want to give to your model.
-* `[SYSTEM_OUTPUT]` is the file that you want to upload.
+* `[SYSTEM_OUTPUT]` is the file that you want to evaluate.
 * `[FILE_TYPE]` is the type of the file, "text", "tsv", "csv", "conll", or "json".
-* `[DATASET]`, `[SUB_DATASET]` and `[SPLIT]` indicate which dataset you're uploading
+* `[DATASET]`, `[SUB_DATASET]` and `[SPLIT]` indicate which dataset you're evaluating
   a system output for.
 * `[SOURCE]` and `[TARGET]` language indicate the language code of the input and output of
   the system. Please refer to the [ISO-639-3](https://iso639-3.sil.org/code_tables/639/data) list for the 3-character 693-3 language codes. Enter `other-[your custom languages]` if the dataset uses custom languages. Enter `none` if the dataset uses other modalities like images. If the inputs and outputs are the in the same language you only need to
@@ -60,11 +63,11 @@ You will need to fill in all the settings appropriately, for example:
 * By default your systems will be private, but if you add the `--public` flag they
   will be made public on the public leaderboards and system listing.
 
-**Uploading w/ Custom Datasets:** You can also upload results for custom datasets that 
-are not supported by DataLab yet:
+**Evaluating w/ Custom Datasets:** You can also evaluate results for custom datasets
+that are not supported by DataLab yet:
 
 ```
-python -m explainaboard_client.cli.upload_system \
+python -m explainaboard_client.cli.evaluate_system \
   --email $EB_EMAIL --api_key $EB_API_KEY \
   --task [TASK_ID] \
   --system_name [MODEL_NAME] \
@@ -78,7 +81,7 @@ interested in getting your datasets directly supported within ExplainaBoard, ple
 open an issue or send a PR to [DataLab](https://github.com/expressai/datalab), and we'll
 be happy to help out!
 
-**Finding Uploaded Systems:** You can also find systems that have already been uploaded 
+**Finding Uploaded Systems:** You can also find systems that have already been evaluated
 using the following syntax
 ```
 python -m explainaboard_client.cli.find_systems \
@@ -100,13 +103,13 @@ Here the `system_ids` are the unique identifier of each system returned in the
 `system_id` field of the JSON returned by the `find_systems` command above. The system
 IDs are *not* the system name as displayed in the interface.
 
-### Uploading Systems to Benchmarks from the Command Line
-Instead of simply uploading an individual system, another common scenario is 
+### Evaluating Systems on Benchmarks from the Command Line
+Instead of simply evaluating an individual system, another common scenario is 
 to submit a group of systems to a benchmark (e.g., GLUE). To achieve this goal,
 you can follow the command below: 
 
 ```shell
-python -m explainaboard_client.cli.upload_benchmark \
+python -m explainaboard_client.cli.evaluate_benchmark \
       --email XXX  \
       --api_key YYY \
       --system_name your_system \
